backend/pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["Maxim Kim <exynil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-rest-framework = "^0.1.0"
djoser = "^2.2.3"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
pylint = "^3.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.format]
max-line-length = 120

[tool.pylint."MESSAGES CONTROL"]
disable = [
  "C0114",  # missing-module-docstring
  "C0116",  # missing-function-docstring
  "R0903",  # too-few-public-methods (for sqlalchemy tables and pydantic classes)
  "C0115",  # missing-class-docstring (for api routes)
  "C0103",  # invalid-name (for pydatic settings)
  ""
]
